+++ src/hotspot/share/classfile/classLoaderExt.cpp
@@ -25,23 +25,18 @@
 #include "precompiled.hpp"
 #include "cds/filemap.hpp"
 #include "classfile/classFileParser.hpp"
-#include "classfile/classFileStream.hpp"
 #include "classfile/classLoader.inline.hpp"
 #include "classfile/classLoaderExt.hpp"
 #include "classfile/classLoaderData.inline.hpp"
 #include "classfile/classLoadInfo.hpp"
-#include "classfile/javaClasses.inline.hpp"
 #include "classfile/klassFactory.hpp"
 #include "classfile/modules.hpp"
-#include "classfile/symbolTable.hpp"
 #include "classfile/systemDictionary.hpp"
-#include "classfile/systemDictionaryShared.hpp"
 #include "classfile/vmSymbols.hpp"
 #include "gc/shared/collectedHeap.hpp"
 #include "logging/log.hpp"
 #include "memory/allocation.inline.hpp"
 #include "memory/resourceArea.hpp"
-#include "memory/oopFactory.hpp"
 #include "oops/instanceKlass.hpp"
 #include "oops/klass.inline.hpp"
 #include "oops/oop.inline.hpp"
@@ -49,9 +44,7 @@
 #include "runtime/arguments.hpp"
 #include "runtime/handles.inline.hpp"
 #include "runtime/java.hpp"
-#include "runtime/javaCalls.hpp"
 #include "runtime/os.hpp"
-#include "services/threadService.hpp"
 #include "utilities/stringUtils.hpp"
 
 jshort ClassLoaderExt::_app_class_paths_start_index = ClassLoaderExt::max_classpath_index;
@@ -252,163 +245,4 @@ void ClassLoaderExt::record_result(const s2 classpath_index, InstanceKlass* resu
   }
   result->set_shared_classpath_index(classpath_index);
   result->set_shared_class_loader_type(classloader_type);
-}
-
-// Load the class of the given name from the location given by path. The path is specified by
-// the "source:" in the class list file (see classListParser.cpp), and can be a directory or
-// a JAR file.
-InstanceKlass* ClassLoaderExt::load_class(Symbol* name, const char* path, TRAPS) {
-  assert(name != NULL, "invariant");
-  assert(DumpSharedSpaces, "this function is only used with -Xshare:dump");
-  ResourceMark rm(THREAD);
-  const char* class_name = name->as_C_string();
-  const char* file_name = file_name_for_class_name(class_name,
-                                                   name->utf8_length());
-  assert(file_name != NULL, "invariant");
-
-  // Lookup stream for parsing .class file
-  ClassFileStream* stream = NULL;
-  ClassPathEntry* e = find_classpath_entry_from_cache(THREAD, path);
-  if (e == NULL) {
-    THROW_NULL(vmSymbols::java_lang_ClassNotFoundException());
-  }
-
-  {
-    PerfClassTraceTime vmtimer(perf_sys_class_lookup_time(),
-                               THREAD->get_thread_stat()->perf_timers_addr(),
-                               PerfClassTraceTime::CLASS_LOAD);
-    stream = e->open_stream(THREAD, file_name);
-  }
-
-  if (stream == NULL) {
-    // open_stream could return NULL even when no exception has be thrown (JDK-8263632).
-    THROW_NULL(vmSymbols::java_lang_ClassNotFoundException());
-    return NULL;
-  }
-  stream->set_verify(true);
-
-  InstanceKlass* k = load_class_impl(name, path, CHECK_NULL);
-
-  SystemDictionaryShared::set_shared_class_misc_info(k, (ClassFileStream*)stream);
-
-  return k;
-}
-
-class URLClassLoaderTable : public ResourceHashtable<
-  Symbol*, Handle,
-  7, // prime number
-  ResourceObj::C_HEAP> {};
-
-static URLClassLoaderTable* _url_classloader_table = NULL;
-
-Handle ClassLoaderExt::create_url_classloader(Symbol* path, TRAPS) {
-  ResourceMark rm(THREAD);
-  JavaValue result(T_OBJECT);
-  Handle path_string = java_lang_String::create_from_str(path->as_C_string(), CHECK_NH);
-  JavaCalls::call_static(&result,
-                         vmClasses::jdk_internal_loader_ClassLoaders_klass(),
-                         vmSymbols::toFileURL_name(),
-                         vmSymbols::toFileURL_signature(),
-                         path_string, CHECK_NH);
-  assert(result.get_type() == T_OBJECT, "just checking");
-  oop url_h = result.get_oop();
-  objArrayHandle urls = oopFactory::new_objArray_handle(vmClasses::URL_klass(), 1, CHECK_NH);
-  urls->obj_at_put(0, url_h);
-
-  Handle url_classloader = JavaCalls::construct_new_instance(
-                             vmClasses::URLClassLoader_klass(),
-                             vmSymbols::url_array_classloader_void_signature(),
-                             urls, Handle(), CHECK_NH);
-  return url_classloader;
-}
-
-Handle ClassLoaderExt::create_and_add_url_classloader(Symbol* path, TRAPS) {
-  Handle url_classloader = create_url_classloader(path, CHECK_NH);
-  bool added = _url_classloader_table->put(path, url_classloader);
-  path->increment_refcount();
-  return url_classloader;
-}
-
-Handle ClassLoaderExt::get_url_classloader(Symbol* path, TRAPS) {
-  if (_url_classloader_table == NULL) {
-    _url_classloader_table = new (ResourceObj::C_HEAP, mtClass)URLClassLoaderTable();
-    Handle url_classloader = create_and_add_url_classloader(path, CHECK_NH);
-    return url_classloader;
-  }
-  Handle* url_classloader_ptr = _url_classloader_table->get(path);
-  if (url_classloader_ptr != NULL) {
-    return *url_classloader_ptr;
-  } else {
-    Handle url_classloader = create_and_add_url_classloader(path, CHECK_NH);
-    return url_classloader;
-  }
-}
-
-InstanceKlass* ClassLoaderExt::load_class_impl(Symbol* name, const char* path, TRAPS) {
-  assert(name != NULL, "invariant");
-  assert(DumpSharedSpaces, "this function is only used with -Xshare:dump");
-
-  Symbol* path_symbol = SymbolTable::new_symbol(path);
-  Handle url_classloader = get_url_classloader(path_symbol, CHECK_NULL);
-  Handle ext_class_name = java_lang_String::externalize_classname(name, CHECK_NULL);
-
-  JavaValue result(T_OBJECT);
-  JavaCallArguments args(2);
-  args.set_receiver(url_classloader);
-  args.push_oop(ext_class_name);
-  args.push_int(JNI_FALSE);
-  JavaCalls::call_virtual(&result,
-                          vmClasses::URLClassLoader_klass(),
-                          vmSymbols::loadClass_name(),
-                          vmSymbols::string_boolean_class_signature(),
-                          &args,
-                          CHECK_NULL);
-  assert(result.get_type() == T_OBJECT, "just checking");
-  oop obj = result.get_oop();
-  InstanceKlass* k = InstanceKlass::cast(java_lang_Class::as_Klass(obj));
-
-  return k;
-}
-
-struct CachedClassPathEntry {
-  const char* _path;
-  ClassPathEntry* _entry;
-};
-
-static GrowableArray<CachedClassPathEntry>* cached_path_entries = NULL;
-
-ClassPathEntry* ClassLoaderExt::find_classpath_entry_from_cache(JavaThread* current, const char* path) {
-  // This is called from dump time so it's single threaded and there's no need for a lock.
-  assert(DumpSharedSpaces, "this function is only used with -Xshare:dump");
-  if (cached_path_entries == NULL) {
-    cached_path_entries = new (ResourceObj::C_HEAP, mtClass) GrowableArray<CachedClassPathEntry>(20, mtClass);
-  }
-  CachedClassPathEntry ccpe;
-  for (int i=0; i<cached_path_entries->length(); i++) {
-    ccpe = cached_path_entries->at(i);
-    if (strcmp(ccpe._path, path) == 0) {
-      if (i != 0) {
-        // Put recent entries at the beginning to speed up searches.
-        cached_path_entries->remove_at(i);
-        cached_path_entries->insert_before(0, ccpe);
-      }
-      return ccpe._entry;
-    }
-  }
-
-  struct stat st;
-  if (os::stat(path, &st) != 0) {
-    // File or directory not found
-    return NULL;
-  }
-  ClassPathEntry* new_entry = NULL;
-
-  new_entry = create_class_path_entry(current, path, &st, false, false);
-  if (new_entry == NULL) {
-    return NULL;
-  }
-  ccpe._path = strdup(path);
-  ccpe._entry = new_entry;
-  cached_path_entries->insert_before(0, ccpe);
-  return new_entry;
 }
